+++ src/olorenchemengine/representations.py
@@ -255,256 +255,6 @@ class TorchGeometricGraph(BaseRepresentation):
         self.atom_featurizer._load(d["atom_featurizer"])
         self.bond_featurizer._load(d["bond_featurizer"])
 
-
-class BaseVecRepresentation(BaseRepresentation):
-    """Representation where given input data, returns a vector representation for each compound."""
-
-    @log_arguments
-    def __init__(
-        self,
-        *args,
-        collinear_thresh=1.01,
-        scale=StandardScaler(),
-        names=None,
-        log=True,
-        **kwargs,
-    ):
-        self.collinear_thresh = collinear_thresh
-        self.to_drop = None
-        if not scale is None:
-            scale = scale.copy()
-        self.scale = scale
-        self.scale_fitted = False
-        self._names = names
-
-        from os import path
-
-        if not path.exists(
-            path.join(
-                oce.CONFIG["CACHE_PATH"], f"vecrep/{self.__class__.__name__}/"
-            )
-        ):
-            os.mkdir(
-                path.join(
-                oce.CONFIG["CACHE_PATH"], f"vecrep/{self.__class__.__name__}/"
-                )
-            )
-
-        super().__init__(*args, log=False, **kwargs)
-
-    @property
-    def names(self):
-        if not self._names is None:
-            return self._names
-        else:
-            raise ValueError(
-                f"Names not set for representation {self.__class__.__name__}"
-            )
-
-    def convert(
-        self,
-        Xs: Union[list, pd.DataFrame, dict, str],
-        ys: Union[list, pd.Series, np.ndarray] = None,
-        fit=False,
-        **kwargs,
-    ) -> List[np.ndarray]:
-        """BaseVecRepresentation's convert returns a list of numpy arrays.
-
-        Args:
-            Xs (Union[list, pd.DataFrame, dict, str]): input data
-            ys (Union[list, pd.Series, np.ndarray], optional): included for compatibility, unused argument. Defaults to None.
-
-        Returns:
-            List[np.ndarray]: list of molecular vector representations
-        """
-        import joblib
-
-        input_hash = (
-            joblib.hash(Xs)
-            + joblib.hash(ys)
-            + joblib.hash(self._save())
-            + joblib.hash(oce.parameterize(self))
-        )
-
-        from os import path
-
-        if oce.CONFIG["CACHE"] and path.exists(
-            path.join(
-                oce.CONFIG["CACHE_PATH"],
-                f"vecrep/{self.__class__.__name__}/{input_hash}.npy",
-            )
-        ):
-            return np.load(
-                path.join(
-                    oce.CONFIG["CACHE_PATH"],
-                    f"vecrep/{self.__class__.__name__}/{input_hash}.npy",
-                ),
-                allow_pickle=True,
-            )
-
-        feats = super().convert(Xs, ys)
-        import pandas as pd
-
-        feats = pd.DataFrame.from_records(
-            feats, columns=[f"col{i}" for i in range(len(feats[0]))]
-        )
-        if fit and len(Xs) > 2:
-            # collinear
-            corr_matrix = feats.corr().abs()
-            upper = corr_matrix.where(
-                np.triu(np.ones(corr_matrix.shape), k=1).astype(np.bool)
-            )
-            self.to_drop = [
-                column
-                for column in upper.columns
-                if any(upper[column] > self.collinear_thresh)
-            ]
-            feats = feats.drop(columns=self.to_drop)
-            # scale
-            if not self.scale is None:
-                x = self.scale.fit_transform(feats.values)
-                feats.values[:] = np.nan_to_num(x.reshape(feats.values.shape))
-                self.scale_fitted = True
-        else:
-            if not self.to_drop is None:
-                feats = feats.drop(columns=self.to_drop)
-            # scale
-            if self.scale_fitted and self.scale is not None:
-                x = self.scale.transform(feats.values)
-                feats.values[:] = np.nan_to_num(x.reshape(feats.values.shape))
-
-        output = np.nan_to_num(np.array(feats.to_records(index=False).tolist()))
-        if oce.CONFIG["CACHE"]:
-            np.save(
-                path.join(
-                    oce.CONFIG["CACHE_PATH"],
-                    f"vecrep/{self.__class__.__name__}/{input_hash}.npy",
-                ),
-                output,
-                allow_pickle=True,
-            )
-        return output
-
-    def calculate_distance(
-        self,
-        x1: Union[str, List[str]],
-        x2: Union[str, List[str]],
-        metric: str = "cosine",
-        **kwargs,
-    ) -> np.ndarray:
-        """Calculates the distance between two molecules or list of molecules.
-
-        Returns a 2D array of distances between each pair of molecules of shape
-        len(x1) by len(x2).
-
-        This uses pairwise_distances from sklearn.metrics to calculate distances
-        between the vector representations of the molecules. Options for distances
-        are Valid values for metric are:
-
-            From scikit-learn: [‘cityblock’, ‘cosine’, ‘euclidean’, ‘l1’, ‘l2’,
-                ‘manhattan’]. These metrics support sparse matrix inputs.
-                [‘nan_euclidean’] but it does not yet support sparse matrices.
-            From scipy.spatial.distance: [‘braycurtis’, ‘canberra’, ‘chebyshev’,
-                ‘correlation’, ‘dice’, ‘hamming’, ‘jaccard’, ‘kulsinski’,
-                ‘mahalanobis’, ‘minkowski’, ‘rogerstanimoto’, ‘russellrao’,
-                ‘seuclidean’, ‘sokalmichener’, ‘sokalsneath’, ‘sqeuclidean’,
-                ‘yule’].
-
-            See the documentation for scipy.spatial.distance for details on these metrics.
-        """
-
-        from sklearn.metrics import pairwise_distances
-
-        if isinstance(x1, str):
-            x1 = [x1]
-        if isinstance(x2, str):
-            x2 = [x2]
-        x1 = self.convert(x1)
-        x2 = self.convert(x2)
-        return pairwise_distances(x1, x2, metric=metric, **kwargs)
-
-    def __add__(self, other):
-        """Adds two representations together
-
-        Parameters:
-            other (BaseVecRepresentation): representation to add to the current representation
-
-        Returns:
-            BaseVecRepresentation: new representation"""
-        return ConcatenatedVecRepresentation(self, other)
-
-    def _save(self):
-        d = {}
-        d.update({"collinear_thresh": self.collinear_thresh})
-        d.update({"to_drop": self.to_drop})
-        if not self.scale is None:
-            d.update({"scale": self.scale._save()})
-            d.update({"scale_fitted": self.scale_fitted})
-        return d
-
-    def _load(self, d):
-        self.collinear_thresh = d["collinear_thresh"]
-        self.to_drop = d["to_drop"]
-        if not self.to_drop is None:
-            self.to_drop = set(self.to_drop)
-        if not self.scale is None and "scale" in d:
-            self.scale._load(d["scale"])
-            self.scale_fitted = d["scale_fitted"]
-
-
-class ConcatenatedVecRepresentation(BaseVecRepresentation):
-    """Creates a structure vector representation by concatenating multiple representations.
-
-    Parameters:
-        rep1 (BaseVecRepresentation): first representation to concatenate
-        rep2 (BaseVecRepresentation): second representation to concatenate
-        log (bool): whether to log the representations or not
-
-    Can be created by adding two representations together using + operator.
-
-    Example
-    ------------------------------
-    import olorenchemengine as oce
-    combo_rep = oce.MorganVecRepresentation(radius=2, nbits=2048) + oce.Mol2Vec()
-    model = oce.RandomForestModel(representation = combo_rep, n_estimators = 1000)
-
-    model.fit(train['Drug'], train['Y'])
-    model.predict(test['Drug'])
-    ------------------------------
-    """
-
-    @log_arguments
-    def __init__(
-        self,
-        rep1: BaseVecRepresentation,
-        rep2: BaseVecRepresentation,
-        log=True,
-        **kwargs,
-    ):
-        self.rep1 = rep1
-        self.rep2 = rep2
-        try:
-            self._names = rep1.names + rep2.names
-        except Exception as e:
-            self._names = None
-        super().__init__(names=self._names, log=False, **kwargs)
-
-    def _convert(self, smiles, y=None, fit=False):
-        converted_1 = self.rep1._convert(smiles, y=y, fit=fit)
-        converted_2 = self.rep2._convert(smiles, y=y, fit=fit)
-        return np.concatenate((converted_1, converted_2))
-
-    def _convert_list(self, smiles_list, ys=None, fit=False):
-        converted_1 = self.rep1._convert_list(smiles_list, ys=ys, fit=fit)
-        converted_2 = self.rep2._convert_list(smiles_list, ys=ys, fit=fit)
-        return np.concatenate((converted_1, converted_2), axis=1)
-
-    def convert(self, smiles_list, ys=None, fit=False, **kwargs):
-        converted_1 = self.rep1.convert(smiles_list, ys=ys, fit=fit)
-        converted_2 = self.rep2.convert(smiles_list, ys=ys, fit=fit)
-        return np.concatenate((converted_1, converted_2), axis=1)
-
-
 class NoisyVec(BaseVecRepresentation):
     """Adds noise to a given BaseVecRepresentation
